Struct wasmer_engine_staticlib::artifact::StaticlibArtifact

pub struct StaticlibArtifact {
    metadata: ModuleMetadata,
    module_bytes: Vec<u8>,
    finished_functions: BoxedSlice<LocalFunctionIndex, FunctionBodyPtr>,
    finished_function_call_trampolines: BoxedSlice<SignatureIndex, VMTrampoline>,
    finished_dynamic_function_trampolines: BoxedSlice<FunctionIndex, FunctionBodyPtr>,
    signatures: BoxedSlice<SignatureIndex, VMSharedSignatureIndex>,
    func_data_registry: Arc<FuncDataRegistry>,
    metadata_length: usize,
    symbol_registry: ModuleMetadataSymbolRegistry,
}

A compiled wasm module, ready to be instantiated.

Fields
metadata: ModuleMetadatamodule_bytes: Vec<u8>finished_functions: BoxedSlice<LocalFunctionIndex, FunctionBodyPtr>finished_function_call_trampolines: BoxedSlice<SignatureIndex, VMTrampoline>finished_dynamic_function_trampolines: BoxedSlice<FunctionIndex, FunctionBodyPtr>signatures: BoxedSlice<SignatureIndex, VMSharedSignatureIndex>func_data_registry: Arc<FuncDataRegistry>metadata_length: usize

Length of the serialized metadata

symbol_registry: ModuleMetadataSymbolRegistry
Implementations

impl StaticlibArtifact

const MAGIC_HEADER_MH_CIGAM_64: &'static [u8]

const MAGIC_HEADER_ELF_32: &'static [u8]

const MAGIC_HEADER_ELF_64: &'static [u8]

const MAGIC_HEADER_COFF_64: &'static [u8]

pub fn is_deserializable(bytes: &[u8]) -> bool

Check if the provided bytes look like StaticlibArtifact.

This means, if the bytes look like a static object file in the target system.

fn generate_metadata<'data>(
data: &'data [u8],
features: &Features,
compiler: &dyn Compiler,
tunables: &dyn Tunables
) -> Result<(CompileModuleInfo, PrimaryMap<LocalFunctionIndex, FunctionBodyData<'data>>, Vec<DataInitializer<'data>>, Option<ModuleTranslationState>), CompileError>

Generate a compilation

pub fn new(
engine: &StaticlibEngine,
data: &[u8],
tunables: &dyn Tunables
) -> Result<Self, CompileError>

Compile a data buffer into a StaticlibArtifact, which can be statically linked against and run later.

pub fn get_default_extension(triple: &Triple) -> &'static str

Get the default extension when serializing this artifact

pub fn from_parts_crosscompiled(
engine_inner: &mut StaticlibEngineInner,
metadata: ModuleMetadata,
module_bytes: Vec<u8>,
metadata_length: usize
) -> Result<Self, CompileError>

Construct a StaticlibArtifact from component parts.

pub unsafe fn deserialize(
engine: &StaticlibEngine,
bytes: &[u8]
) -> Result<Self, DeserializeError>

Deserialize a StaticlibArtifact from bytes.

Safety

The bytes must represent a serialized WebAssembly module.

pub fn symbol_registry(&self) -> &dyn SymbolRegistry

Get the SymbolRegistry used to generate the names used in the Artifact.

pub fn metadata_length(&self) -> usize

The length in bytes of the metadata in the serialized output.