Struct wasmer_engine_staticlib::engine::StaticlibEngineInner

pub struct StaticlibEngineInner {
    compiler: Option<Box<dyn Compiler>>,
    features: Features,
    signatures: SignatureRegistry,
    func_data: Arc<FuncDataRegistry>,
    prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>,
}

The inner contents of StaticlibEngine

Fields
compiler: Option<Box<dyn Compiler>>

The compiler

features: Features

The WebAssembly features to use

signatures: SignatureRegistry

The signature registry is used mainly to operate with trampolines performantly.

func_data: Arc<FuncDataRegistry>

The backing storage of VMFuncRefs. This centralized store ensures that 2 functions with the same VMCallerCheckedAnyfunc will have the same VMFuncRef. It also guarantees that the VMFuncRefs stay valid until the engine is dropped.

prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>

The prefixer returns the a String to prefix each of the functions in the static object generated by the StaticlibEngine, so we can assure no collisions.

Implementations

impl StaticlibEngineInner

pub fn compiler(&self) -> Result<&dyn Compiler, CompileError>

Gets the compiler associated to this engine.

pub(crate) fn get_prefix(&self, bytes: &[u8]) -> String

pub(crate) fn features(&self) -> &Features

pub fn validate<'data>(&self, data: &'data [u8]) -> Result<(), CompileError>

Validate the module

pub fn signatures(&self) -> &SignatureRegistry

Shared signature registry.

pub(crate) fn func_data(&self) -> &Arc<FuncDataRegistry>

Shared func metadata registry.